Restructured charm, added layer-service-account, added options force-user, force-group, force-create-mask, force-dir-mask, dir-mask, write-list, WIP smb-users option
