--- build.sh
@@ -1,2 +1,0 @@
-git describe --tags > VERSION
-charm build . --force
