--- config.yaml
@@ -1,26 +1,0 @@
-options:
-  smb-shares:
-    type: string
-    default: ""
-    description: "Comma seperated list of folders to share"
-  smb-browsable:
-    type: boolean
-    default: True
-    description: "Shares are browsable"
-  smb-guest:
-    type: boolean
-    default: True
-    description: "Shares allow guest access"
-  smb-read-only:
-    type: boolean
-    default: False
-    description: "Shares are read only"
-  smb-create-mask:
-    type: string
-    default: "0777"
-    description: "Smb.conf reate mask setting for shares"
-  smb-custom:
-    type: string
-    default: ""
-    description: "Raw configuration provide key=value comma seperated keys semicolon seperated shares"
-
