--- layer.yaml
@@ -1,2 +1,0 @@
-includes: ['layer:basic', 'layer:version']  # if you use any interfaces, add them here
-repo: git@github.com:chris-sanders/layer-samba-server.git
