--- metadata.yaml
@@ -1,21 +1,0 @@
-name: samba-server
-summary: Samba file server
-maintainer: Chris Sanders <sanders.chris@gmail.com>
-description: |
-  Installs Samba file server and makes local folders avaialble.
-series:
-  - xenial
-tags:
-  # Replace "misc" with one or more whitelisted tags from this list:
-  # https://jujucharms.com/docs/stable/authors-charm-metadata
-  - storage 
-subordinate: false
-# provides:
-#   provides-relation:
-#     interface: interface-name
-# requires:
-#   requires-relation:
-#     interface: interface-name
-# peers:
-#   peer-relation:
-#     interface: interface-name
