+++ src/templates/smb.conf
@@ -1,0 +1,15 @@
+# This file is managed by the samba-server charm - edits will not persist
+
+{% if smb_shares -%}
+{% for share in smb_shares.split(',') -%}
+{% for client in smb_clients.split(',') -%}
+{{ share }}	{{ client }}({{ smb_options }})
+{% endfor -%}
+{% endfor -%}
+{% endif -%}
+
+{% if smb_custom -%}
+{% for line in smb_custom.split(';') -%}
+{{ line }}
+{% endfor -%}
+{% endif -%}
