--- templates/smb.conf
@@ -1,15 +1,0 @@
-# This file is managed by the nfs-server charm edits will not persist
-
-{% if nfs_shares -%}
-{% for share in nfs_shares.split(',') -%}
-{% for client in nfs_clients.split(',') -%}
-{{ share }}	{{ client }}({{ nfs_options }})
-{% endfor -%}
-{% endfor -%}
-{% endif -%}
-
-{% if nfs_custom -%}
-{% for line in nfs_custom.split(';') -%}
-{{ line }}
-{% endfor -%}
-{% endif -%}
